+++ app/src/main/res/layout/activity_notification_webview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NotificationWebviewActivity">
+    tools:context=".activities.Notifications.NotificationWebviewActivity">
 
     <WebView
         android:layout_width="match_parent"
